Separate preprocessing and solving into separate modules
Separate preprocessing and solving into separate modules.

* Add `src/preprocessing.rs` to contain preprocessing functions.
* Move `repeatedly_resolve_and_update` function from `src/cnf.rs` to `src/preprocessing.rs`.
* Add `save_transformed_cnf` function in `src/preprocessing.rs` to save the transformed CNF formula to a new file.
* Remove `repeatedly_resolve_and_update` function from `src/cnf.rs`.
* Import `preprocessing` module in `src/cnf.rs`.
* Update `solve` function in `src/main.rs` to call preprocessing functions from `src/preprocessing.rs`.
* Add an option in `src/main.rs` to save the transformed CNF formula to a new file.
* Import `preprocessing` module in `src/main.rs`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/AHartNtkn/odesat?shareId=XXXX-XXXX-XXXX-XXXX).
